refactor: crypto and stark related functions
