+++ imx-core-sdk-android/src/main/java/com/immutable/sdk/workflows/BuyCryptoWorkflow.kt
@@ -95,7 +95,9 @@ internal fun buyCrypto(
             url = url,
             toolbarColourInt = colourInt
         )
-    } catch (e: Exception) { e.cause?.let { throw it } ?: throw e }
+    } catch (e: Exception) {
+        e.cause?.let { throw it } ?: throw e
+    }
 }
 
 @Suppress("TooGenericExceptionCaught", "InstanceOfCheckForException")
